% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{Cookie}
\alias{Cookie}
\title{Create a Cookie}
\description{
Create a cookie object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Cookie$new()}}
\item \href{#method-get}{\code{Cookie$get()}}
\item \href{#method-set}{\code{Cookie$set()}}
\item \href{#method-rm}{\code{Cookie$rm()}}
\item \href{#method-clone}{\code{Cookie$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cookie$new(name, session = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the cookie.}

\item{\code{session}}{A valid Shiny session.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a Cookie
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cookie$get()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Get the value of the cookie
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cookie$set(value, expires = 365L, path = "/")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{Value of the cookie}

\item{\code{expires}}{When the cookie is set to expire, integer indicating number of days.}

\item{\code{path}}{Path where cookie is visible.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Set the value of the cookie
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rm"></a>}}
\subsection{Method \code{rm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cookie$rm()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Remove the cookie.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Cookie$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
