% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cookies.R
\name{fetch_cookies}
\alias{fetch_cookies}
\alias{fetch_cookie}
\alias{add_cookie}
\alias{remove_cookie}
\title{Add and fetch browsers cookies}
\usage{
fetch_cookies(input, session)

fetch_cookie(name, input, session)

add_cookie(name, value, session)

remove_cookie(name, session)
}
\arguments{
\item{input, session}{the `input` and `session` object from Shiny}

\item{name}{name of the cookie to set/fetch}

\item{value}{the value to set for the cookie}
}
\description{
`fetch_cookies` returns all the cookies, while `fetch_cookie` search for
one cookie in the browser.
}
